src/requirements.txt
altair==4.2.2
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
blinker==1.5
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cramjam==2.6.2
cycler==0.11.0
Cython==0.29.33
debugpy==1.5.1
decorator==5.1.1
distlib==0.3.6
entrypoints==0.4
executing==1.2.0
fastparquet==2023.2.0
filelock==3.9.0
fonttools==4.38.0
fsspec==2023.1.0
fuzzywuzzy==0.18.0
gitdb==4.0.10
GitPython==3.1.30
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.15.0
ipython==8.10.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==8.0.2
jupyter_core==5.2.0
kiwisolver==1.4.4
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.6
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
pip==22.3.1
pipenv==2023.2.4
platformdirs==3.0.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.0
pure-eval==0.2.2
pyarrow==11.0.0
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pywin32==305.1
pyzmq==23.2.0
requests==2.28.2
rich==13.3.1
scikit-learn==1.2.1
scikit-surprise==1.1.3
scipy==1.10.0
semver==2.13.0
setuptools==67.2.0
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
streamlit==1.18.1
streamlit-option-menu==0.3.2
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
tornado==6.2
traitlets==5.9.0
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
validators==0.20.0
virtualenv==20.19.0
virtualenv-clone==0.5.7
watchdog==2.2.1
wcwidth==0.2.6
wheel==0.38.4
wincertstore==0.2
zipfile36==0.1.3
zipp==3.13.0
